@@ -1,14 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
+using MediatR;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using System;
-using System.Collections.Generic;
-using MediatR;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
 using SFA.DAS.Api.Common.AppStart;
 using SFA.DAS.Api.Common.Configuration;
@@ -17,98 +16,85 @@ using SFA.DAS.Recruit.Application.Queries.GetTrainingProgrammes;
 using SFA.DAS.SharedOuterApi.AppStart;
 using SFA.DAS.SharedOuterApi.Infrastructure.HealthCheck;
 
-namespace SFA.DAS.Recruit.Api
+namespace SFA.DAS.Recruit.Api;
+
+public static class Startup
 {
-    public class Startup
+    public static void ConfigureServices(
+        IServiceCollection services, 
+        IWebHostEnvironment environment,
+        IConfigurationRoot configuration)
     {
-        private readonly IWebHostEnvironment _env;
-        private readonly IConfiguration _configuration;
+        services.AddSingleton(environment);
+        services.AddConfigurationOptions(configuration);
 
-        public Startup(IConfiguration configuration, IWebHostEnvironment env)
+        if (!configuration.IsLocalOrDev())
         {
-            _env = env;
-            _configuration = configuration.BuildSharedConfiguration();
+            var azureAdConfiguration = configuration
+                .GetSection("AzureAd")
+                .Get<AzureActiveDirectoryConfiguration>();
+            var policies = new Dictionary<string, string>
+            {
+                {"default", "APIM"}
+            };
+
+            services.AddAuthentication(azureAdConfiguration, policies);
         }
 
-        public void ConfigureServices(IServiceCollection services)
+        services.AddMediatR(typeof(GetTrainingProgrammesQuery).Assembly);
+        services.AddServiceRegistration();
+
+        services.Configure<RouteOptions>(options =>
+            {
+                options.LowercaseUrls = true;
+                options.LowercaseQueryStrings = true;
+            }).AddMvc(o =>
+            {
+                if (!configuration.IsLocalOrDev())
+                {
+                    o.Filters.Add(new AuthorizeFilter("default"));
+                }
+            }).AddJsonOptions(options => options.JsonSerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull);
+
+        if (configuration["Environment"] != "DEV")
         {
-            services.AddSingleton(_env);
-
-            services.AddConfigurationOptions(_configuration);
-
-            if (!_configuration.IsLocalOrDev())
-            {
-                var azureAdConfiguration = _configuration
-                    .GetSection("AzureAd")
-                    .Get<AzureActiveDirectoryConfiguration>();
-                var policies = new Dictionary<string, string>
-                {
-                    {"default", "APIM"}
-                };
-
-                services.AddAuthentication(azureAdConfiguration, policies);
-            }
-
-            services.AddMediatR(typeof(GetTrainingProgrammesQuery).Assembly);
-            services.AddServiceRegistration();
-
-            services.Configure<RouteOptions>(options =>
-                {
-                    options.LowercaseUrls = true;
-                    options.LowercaseQueryStrings = true;
-                }).AddMvc(o =>
-                {
-                    if (!_configuration.IsLocalOrDev())
-                    {
-                        o.Filters.Add(new AuthorizeFilter("default"));
-                    }
-                }).SetCompatibilityVersion(CompatibilityVersion.Version_3_0)
-                .AddJsonOptions(options => options.JsonSerializerOptions.IgnoreNullValues = true);
-
-            if (_configuration["Environment"] != "DEV")
-            {
-                services.AddHealthChecks()
-                    .AddCheck<CoursesApiHealthCheck>("Courses API health check")
-                    .AddCheck<CourseDeliveryApiHealthCheck>("Course Delivery API health check");
-            }
-            
-            services.AddApplicationInsightsTelemetry(_configuration["APPINSIGHTS_INSTRUMENTATIONKEY"]);
-
-            services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new OpenApiInfo { Title = "RecruitOuterApi", Version = "v1" });
-            });
+            services.AddHealthChecks()
+                .AddCheck<CoursesApiHealthCheck>("Courses API health check")
+                .AddCheck<CourseDeliveryApiHealthCheck>("Course Delivery API health check");
         }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseAuthentication();
-
-            if (!_configuration["Environment"].Equals("DEV", StringComparison.CurrentCultureIgnoreCase))
-            {
-                app.UseHealthChecks();
-            }
-            
-            app.UseRouting();
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllerRoute(
-                    name: "default",
-                    pattern: "api/{controller=Courses}/{action=index}/{id?}");
-            });
         
-            app.UseSwagger();
-            app.UseSwaggerUI(c =>
-            {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "RecruitOuterApi");
-                c.RoutePrefix = string.Empty;
-            });
+        //todo obsolete, should use connectionstring instead https://github.com/microsoft/ApplicationInsights-dotnet/issues/2560
+        //services.AddApplicationInsightsTelemetry(options => options.ConnectionString = configuration.GetConnectionString("ApplicationInsights"));
+        services.AddApplicationInsightsTelemetry(configuration["APPINSIGHTS_INSTRUMENTATIONKEY"]);
+
+        services.AddSwaggerGen(c =>
+        {
+            c.SwaggerDoc("v1", new OpenApiInfo { Title = "RecruitOuterApi", Version = "v1" });
+        });
+    }
+    
+    public static void ConfigureApp(IApplicationBuilder app, IConfigurationRoot configuration)
+    {
+        app.UseAuthentication();
+
+        if (!configuration["Environment"].Equals("DEV", StringComparison.CurrentCultureIgnoreCase))
+        {
+            app.UseHealthChecks();
         }
+                
+        app.UseRouting();
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllerRoute(
+                name: "default",
+                pattern: "api/{controller=Providers}/{action=index}/{id?}");
+        });
+            
+        app.UseSwagger();
+        app.UseSwaggerUI(c =>
+        {
+            c.SwaggerEndpoint("/swagger/v1/swagger.json", "RecruitOuterApi");
+            c.RoutePrefix = string.Empty;
+        });
     }
 }
